chore: add reasoning stable v
